+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.12.1
+
+### Added
+
+- REDCap PID 879: `{instrument}_complete` → `"Curious Track.{instrument}_received".
+
+### Deprecated
+
+- `hbnmigration.from_redcap.from_redcap` `export_fields` parameter.
+
 ## 1.12.0
 
 ### Changed